% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_hgmm}
\alias{download_hgmm}
\title{General download function}
\usage{
download_hgmm(file_path, dataset = "hgmm100", force = FALSE)
}
\arguments{
\item{file_path}{Character vector of length 1, specifying where to download
the data.}

\item{dataset}{Character, should be either \code{hgmm100} (for the 100 cell
dataset) or \code{hgmm1k} (for the 1k cell dataset).}

\item{force}{Logical, whether to force redownload if the files are already
present. Defaults to \code{FALSE}.}
}
\value{
Character, directory to be used in \code{BUSpaRse}.
}
\description{
This function will download the 10x 100 1:1 Mixture of Fresh Frozen Human
(HEK293T) and Mouse (NIH3T3) Cells dataset, already processed and stored in
the BUS format, from \code{ExperimentHub}. This function will decompress the
downloaded file and return the directory where the files necessary to
construct the sparse matrix with \code{BUSpaRse} are located.
}
\examples{
download_hgmm(".", dataset = "hgmm100")

}
