% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\docType{package}
\name{TENxhgmmBUS}
\alias{TENxhgmmBUS}
\alias{TENxhgmmBUS-package}
\title{TENxhgmmBUS: 10x human and mouse cell mixture datasets in the BUS format}
\description{
This package provides 1:1 Mixture of Fresh Frozen Human (HEK293T) and Mouse
(NIH3T3) Cells datasets from 10x Genomics (v2 chemistry) in the
\href{https://www.biorxiv.org/content/early/2018/11/21/472571}{BUS format},
to be downloaded from within R. The files downloaded from this package are
sufficient to generate a sparse matrix with package
\href{https://github.com/BUStools/BUSpaRse}{BUSpaRse}
to be used for downstream analysis with Seurat. Both datasets provided are
of the version of November 8, 2017.
}
\section{Datasets}{

\describe{
\item{hgmm100}{The 100 cell dataset. The raw data can be found here:
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/hgmm_100}}
\item{hgmm1k}{The 1000 cell dataset. The raw data can be found here:
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/hgmm_1k}}
}
}

